requirements.txt
alembic==1.14.0
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_JWT_Extended==4.7.0
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.36
